Move sfr_kernels module to fitting_helpers/fitting_kernels.py and deleted redundant functionality
