+++ diffstar/fitting_helpers/stars.py
@@ -1,226 +1,9 @@
 """
 """
 import numpy as np
-from jax import jit as jjit
 
-from ..kernels import main_sequence_kernels as msk
-from ..kernels import sfr_kernels as sfrk
 from ..utils import _get_dt_array
-
-
-@jjit
-def calculate_sm_sfr_fstar_history_from_mah(
-    lgt,
-    dt,
-    dmhdt,
-    log_mah,
-    u_ms_params,
-    u_q_params,
-    index_select,
-    index_high,
-    fstar_tdelay,
-):
-    """Calculate individual galaxy SFH from precalculated halo MAH
-
-    The accretion rate of gas is proportional to
-    the accretion rate of the halo; main sequence galaxies transform accreted
-    gas into stars over a time depletion timescale with an efficiency that
-    depends on the instantaneous halo mass; some galaxies experience a quenching
-    event and may subsequently experience a rejuvenated star formation.
-
-    Parameters
-    ----------
-    lgt : ndarray of shape (n_times, )
-        Base-10 log of cosmic time of each simulated snapshot in Gyr
-
-    dt : ndarray of shape (n_times, )
-        Cosmic time steps between each simulated snapshot in Gyr
-
-    dmhdt : ndarray of shape (n_times, )
-        Diffmah halo mass accretion rate in units of Msun/yr
-
-    log_mah : ndarray of shape (n_times, )
-        Diffmah halo mass accretion history in units of Msun
-
-    u_ms_params : ndarray of shape (5, )
-        Star formation efficiency model unbounded parameters. Includes
-        (u_lgmcrit, u_lgy_at_mcrit, u_indx_lo, u_indx_hi, u_tau_dep)
-
-    u_q_params : ndarray of shape (4, )
-        Quenching model unbounded parameters. Includes (u_qt, u_qs, u_drop, u_rejuv)
-
-    index_select: ndarray of shape (n_times_fstar, )
-        Snapshot indices used in fstar computation
-
-    index_high: ndarray of shape (n_times_fstar, )
-        Indices of np.searchsorted(t, t - fstar_tdelay)[index_select]
-
-    fstar_tdelay: float
-        Time interval in Gyr for fstar definition.
-        fstar = (mstar(t) - mstar(t-fstar_tdelay)) / fstar_tdelay[Gyr]
-
-    Returns
-    -------
-    mstar : ndarray of shape (n_times)
-        Cumulative stellar mass history in units of Msun assuming h=1
-
-    sfr : ndarray of shape (n_times)
-        Star formation rate history in units of Msun/yr assuming h=1
-
-    fstar : ndarray of shape (n_times)
-        SFH averaged over timescale fstar_tdelay in units of Msun/yr assuming h=1
-
-    """
-    return sfrk.calculate_sm_sfr_fstar_history_from_mah(
-        lgt,
-        dt,
-        dmhdt,
-        log_mah,
-        u_ms_params,
-        u_q_params,
-        index_select,
-        index_high,
-        fstar_tdelay,
-    )
-
-
-@jjit
-def calculate_sm_sfr_history_from_mah(
-    lgt,
-    dt,
-    dmhdt,
-    log_mah,
-    u_ms_params,
-    u_q_params,
-):
-    """Calculate individual galaxy SFH from precalculated halo MAH
-
-    The accretion rate of gas is proportional to
-    the accretion rate of the halo; main sequence galaxies transform accreted
-    gas into stars over a time depletion timescale with an efficiency that
-    depends on the instantaneous halo mass; some galaxies experience a quenching
-    event and may subsequently experience a rejuvenated star formation.
-
-    Parameters
-    ----------
-    lgt : ndarray of shape (n_times, )
-        Base-10 log of cosmic time of each simulated snapshot in Gyr
-
-    dt : ndarray of shape (n_times, )
-        Cosmic time steps between each simulated snapshot in Gyr
-
-    dmhdt : ndarray of shape (n_times, )
-        Diffmah halo mass accretion rate in units of Msun/yr
-
-    log_mah : ndarray of shape (n_times, )
-        Diffmah halo mass accretion history in units of Msun
-
-    u_ms_params : ndarray of shape (5, )
-        Star formation efficiency model unbounded parameters. Includes
-        (u_lgmcrit, u_lgy_at_mcrit, u_indx_lo, u_indx_hi, u_tau_dep)
-
-    u_q_params : ndarray of shape (4, )
-        Quenching model unbounded parameters. Includes (u_qt, u_qs, u_drop, u_rejuv)
-
-    Returns
-    -------
-    mstar : ndarray of shape (n_times)
-        Cumulative stellar mass history in units of Msun assuming h=1
-
-    sfr : ndarray of shape (n_times)
-        Star formation rate history in units of Msun/yr assuming h=1
-
-    """
-    return sfrk.calculate_sm_sfr_history_from_mah(
-        lgt,
-        dt,
-        dmhdt,
-        log_mah,
-        u_ms_params,
-        u_q_params,
-    )
-
-
-@jjit
-def calculate_histories(
-    lgt,
-    dt,
-    mah_params,
-    u_ms_params,
-    u_q_params,
-    index_select,
-    index_high,
-    fstar_tdelay,
-):
-    """Calculate individual halo mass MAH and galaxy SFH
-
-    The accretion rate of gas is proportional to the accretion rate of the halo;
-    main sequence galaxies transform accreted
-    gas into stars over a time depletion timescale with an efficiency that
-    depends on the instantaneous halo mass; some galaxies experience a quenching
-    event and may subsequently experience a rejuvenated star formation.
-
-    Parameters
-    ----------
-    lgt : ndarray of shape (n_times, )
-        Base-10 log of cosmic time of each simulated snapshot in Gyr
-
-    dt : ndarray of shape (n_times, )
-        Cosmic time steps between each simulated snapshot in Gyr
-
-    mah_params : ndarray of shape (6, )
-        Best fit diffmah halo parameters. Includes
-        (logt0, logmp, logtc, k, early, late)
-
-    u_ms_params : ndarray of shape (5, )
-        Star formation efficiency model unbounded parameters. Includes
-        (u_lgmcrit, u_lgy_at_mcrit, u_indx_lo, u_indx_hi, u_tau_dep)
-
-    u_q_params : ndarray of shape (4, )
-        Quenching model unbounded parameters. Includes
-        (u_qt, u_qs, u_drop, u_rejuv)
-
-    index_select: ndarray of shape (n_times_fstar, )
-        Snapshot indices used in fstar computation.
-
-    index_high: ndarray of shape (n_times_fstar, )
-        Indices of np.searchsorted(t, t - fstar_tdelay)[index_select]
-
-    fstar_tdelay: float
-        Time interval in Gyr for fstar definition.
-        fstar = (mstar(t) - mstar(t-fstar_tdelay)) / fstar_tdelay[Gyr]
-
-    Returns
-    -------
-    mstar : ndarray of shape (n_times)
-        Cumulative stellar mass history in units of Msun assuming h=1
-
-    sfr : ndarray of shape (n_times)
-        Star formation rate history in units of Msun/yr assuming h=1
-
-    fstar : ndarray of shape (n_times_fstar)
-        SFH averaged over timescale fstar_tdelay in units of Msun/yr assuming h=1
-
-    dmhdt : ndarray of shape (n_times)
-        Mass accretion rate in units of Msun/yr assuming h=1
-
-    log_mah : ndarray of shape (n_times)
-        Base-10 log of cumulative peak halo mass in units of Msun assuming h=1
-
-    """
-    return sfrk.calculate_histories(
-        lgt,
-        dt,
-        mah_params,
-        u_ms_params,
-        u_q_params,
-        index_select,
-        index_high,
-        fstar_tdelay,
-    )
-
-
-calculate_histories_vmap = sfrk.calculate_histories_vmap
+from .fitting_kernels import calculate_histories_vmap
 
 
 def calculate_histories_batch(t_sim, mah_params, u_ms_params, u_q_params, fstar_tdelay):
@@ -319,163 +102,6 @@ def calculate_histories_batch(t_sim, mah_params, u_ms_params, u_q_params, fstar_
         log_mah[inds] = _res[4]
 
     return mstar, sfr, fstar, dmhdt, log_mah
-
-
-@jjit
-def _get_bounded_sfr_params(
-    u_lgmcrit,
-    u_lgy_at_mcrit,
-    u_indx_lo,
-    u_indx_hi,
-    u_tau_dep,
-):
-    return msk._get_bounded_sfr_params(
-        u_lgmcrit,
-        u_lgy_at_mcrit,
-        u_indx_lo,
-        u_indx_hi,
-        u_tau_dep,
-    )
-
-
-@jjit
-def _get_unbounded_sfr_params(
-    lgmcrit,
-    lgy_at_mcrit,
-    indx_lo,
-    indx_hi,
-    tau_dep,
-):
-    return msk._get_unbounded_sfr_params(
-        lgmcrit,
-        lgy_at_mcrit,
-        indx_lo,
-        indx_hi,
-        tau_dep,
-    )
-
-
-_get_bounded_sfr_params_vmap = msk._get_bounded_sfr_params_vmap
-_get_unbounded_sfr_params_vmap = msk._get_unbounded_sfr_params_vmap
-
-
-@jjit
-def _integrate_sfr(sfr, dt):
-    """Calculate the cumulative stellar mass history."""
-    return sfrk._integrate_sfr(sfr, dt)
-
-
-@jjit
-def compute_fstar(tarr, mstar, index_select, index_high, fstar_tdelay):
-    """Time averaged SFH that has ocurred over some previous time period
-
-    fstar = (mstar(t) - mstar(t-fstar_tdelay)) / fstar_tdelay
-    Parameters
-    ----------
-    tarr : ndarray of shape (n_times, )
-        Cosmic time of each simulated snapshot in Gyr
-
-    mstar : ndarray of shape (n_times, )
-        Stellar mass history in Msun units
-
-    index_select: ndarray of shape (n_times_fstar, )
-        Snapshot indices used in fstar computation
-
-    index_high: ndarray of shape (n_times_fstar, )
-        Indices of np.searchsorted(t, t - fstar_tdelay)[index_select]
-
-    fstar_tdelay: float
-        Time interval in Gyr units for fstar definition.
-        fstar = (mstar(t) - mstar(t-fstar_tdelay)) / fstar_tdelay
-
-    Returns
-    -------
-    fstar : ndarray of shape (n_times)
-        SFH averaged over timescale fstar_tdelay in units of Msun/yr assuming h=1
-
-    """
-    return sfrk.compute_fstar(tarr, mstar, index_select, index_high, fstar_tdelay)
-
-
-compute_fstar_vmap = sfrk.compute_fstar_vmap
-
-
-@jjit
-def _sfr_history_from_mah(lgt, dtarr, dmhdt, log_mah, u_ms_params, u_q_params):
-    """Star formation history of an individual galaxy.
-
-    SFH(t) = Quenching(t) x epsilon(Mhalo) int Depletion(t|t') x Mgas(t') dt'.
-    Mgas(t) = FB x dMhalo(t)/dt
-
-    Parameters
-    ----------
-    lgt : ndarray of shape (n_times, )
-        Base-10 log of cosmic time of each simulated snapshot in Gyr
-
-    dtarr : ndarray of shape (n_times, )
-        Cosmic time steps between each simulated snapshot in Gyr
-
-    dmhdt : ndarray of shape (n_times, )
-        Diffmah halo mass accretion rate in units of Msun/yr
-
-    log_mah : ndarray of shape (n_times, )
-        Diffmah halo mass accretion history in units of Msun
-
-    u_ms_params : ndarray of shape (5, )
-        Star formation efficiency model unbounded parameters. Includes
-        (u_lgmcrit, u_lgy_at_mcrit, u_indx_lo, u_indx_hi, u_tau_dep)
-
-    u_q_params : ndarray of shape (4, )
-        Quenching model unbounded parameters. Includes
-        (u_qt, u_qs, u_drop, u_rejuv)
-
-    Returns
-    -------
-    sfr : ndarray of shape (n_times)
-        Star formation rate history in units of Msun/yr assuming h=1.
-
-    """
-    return sfrk._sfr_history_from_mah(
-        lgt, dtarr, dmhdt, log_mah, u_ms_params, u_q_params
-    )
-
-
-@jjit
-def _ms_sfr_history_from_mah(lgt, dtarr, dmhdt, log_mah, u_ms_params):
-    """Main Sequence formation history of an individual galaxy."""
-
-    return sfrk._ms_sfr_history_from_mah(lgt, dtarr, dmhdt, log_mah, u_ms_params)
-
-
-@jjit
-def _sfr_eff_plaw(lgm, lgmcrit, lgy_at_mcrit, indx_lo, indx_hi):
-    """Instantaneous baryon conversion efficiency of main sequence galaxies
-
-    Main sequence efficiency kernel, epsilon(Mhalo)
-
-    Parameters
-    ----------
-    lgm : ndarray of shape (n_times, )
-        Diffmah halo mass accretion history in units of Msun
-
-    lgmcrit : float
-        Base-10 log of the critical mass
-    lgy_at_mcrit : float
-        Base-10 log of the critical efficiency at critical mass
-
-    indx_lo : float
-        Asymptotic value of the efficiency at low halo masses
-
-    indx_hi : float
-        Asymptotic value of the efficiency at high halo masses
-
-    Returns
-    -------
-    efficiency : ndarray of shape (n_times)
-        Main sequence efficiency value at each snapshot
-
-    """
-    return sfrk._sfr_eff_plaw(lgm, lgmcrit, lgy_at_mcrit, indx_lo, indx_hi)
 
 
 def fstar_tools(t_sim, fstar_tdelay=1.0):
