+++ diffstar/tests/test_diffstar_is_frozen.py
@@ -11,7 +11,7 @@ from diffmah.individual_halo_assembly import (
 from jax import numpy as jnp
 
 from ..defaults import DEFAULT_U_MS_PARAMS, DEFAULT_U_Q_PARAMS, LGT0
-from ..fitting_helpers.stars import _sfr_history_from_mah
+from ..fitting_helpers.fitting_kernels import _sfr_history_from_mah
 from ..kernels.quenching_kernels import _get_unbounded_q_params
 from ..utils import _get_dt_array
